[libs] E: Dont Use size_t in IKC Messages
This commit fixes IKC messages to use u32 instead of size_t to make them
platform independent.

+++ src/libs/syscall/src/unistd/message/pwrite.rs
@@ -31,7 +31,7 @@ use sysapi::sys_types::ssize_t;
 #[repr(C, packed)]
 pub struct PartialWriteRequest {
     pub fd: i32,
-    pub count: size_t,
+    pub count: u32,
     pub offset: off_t,
     pub buffer: [u8; Self::BUFFER_SIZE],
 }
@@ -40,10 +40,10 @@ pub struct PartialWriteRequest {
 impl PartialWriteRequest {
     pub const BUFFER_SIZE: usize = LinuxDaemonMessage::PAYLOAD_SIZE
         - mem::size_of::<i32>()
-        - mem::size_of::<i32>()
+        - mem::size_of::<u32>()
         - mem::size_of::<off_t>();
 
-    fn new(fd: i32, count: size_t, offset: off_t, buffer: [u8; Self::BUFFER_SIZE]) -> Self {
+    fn new(fd: i32, count: u32, offset: off_t, buffer: [u8; Self::BUFFER_SIZE]) -> Self {
         Self {
             fd,
             count,
